chore(*): split lint module into smaller submodules, added tests
