+++ src/lint/index.ts
@@ -1,0 +1,1 @@
+export { lintText, lintFile, lintFolder, lintProject } from './lint'
