+++ src/types/Process.d.ts
@@ -1,0 +1,6 @@
+declare namespace NodeJS {
+  export interface Process {
+    projectDir: string
+    currentDir: string
+  }
+}
